# swagger.yaml
swagger: "2.0"
info:
  title: "API Authentication/authorization with Node/Express/Passport"
  version: "1.0.0"
  description: |
    Step by step authentication and authorization for a Node/Express API. 
    I will use a passport localStrategy and sign Access and refresh tokens with jsonwebtoken.
    You must first create a user (/auth/signup endpoint to get your authorization).
  contact:
    name: GoJrM 
host: "localhost:5000"
basePath: "/api/v1"
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces: 
- "application/json"
paths:
...
  /auth/login:
    post:
      summary: "authenticate users"
      description: "authenticate users and return accessToken/refreshToken for authorization"
      operationId: "userLogIn"
      consumes:
        - "application/json"
        - "application/x-form-urlencoded"
      tags:
        - "Auth"
      parameters:
        - in: "body"
          name: "body"
          description: "users email and password"
          schema:
            type: object
            required:
            - email
            - password            
            properties: 
              password:
                type: "string"
                minimum: 6
              email:
                type: "string"
                minimum: 6
      responses:
        200:
          description: "User is authenticated"
          schema:
            type: object
            properties:
              message:
                type: string
              accessToken:
                type: string
              expiresIn:
                type: number
              created: 
                type: string
                format: date-time
              refreshToken:
                type: string
              timestamp: 
                type: string
...
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "authorization"
    in: "header"
...